处理冲突

+++ services/media_engine/plugins/source/http_source/hls/hls_playlist_downloader.cpp
@@ -353,7 +353,7 @@ void HlsPlayListDownloader::SetMimeType(const std::string& mimeType)
 
 size_t HlsPlayListDownloader::GetSegmentOffset(uint32_t tsIndex)
 {
-    if (master_ && master_->segmentOffsets_.size() > static_cast<size_t>(tsIndex)) {
+    if (master_ && master_->segmentOffsets_.size() > tsIndex) {
         return master_->segmentOffsets_[tsIndex];
     }
     return 0;
